requirements.txt
streamlit
requests
streamlit_lottie
pillow
simplejson
joblib
numpy
pandas
scikit-learn
geopandas
matplotlib